move loading-screen component to common folder
