+++ bigbluebutton-html5/imports/ui/components/common/loading-screen/component.jsx
