+++ bigbluebutton-html5/imports/ui/components/settings/submenus/application/styles.js
@@ -6,7 +6,7 @@ import {
   colorGrayLighter,
 } from '/imports/ui/stylesheets/styled-components/palette';
 import { borderSize, borderSizeLarge } from '/imports/ui/stylesheets/styled-components/general';
-import SpinnerStyles from '/imports/ui/components/loading-screen/styles';
+import SpinnerStyles from '/imports/ui/components/common/loading-screen/styles';
 import Styled from '/imports/ui/components/settings/submenus/styles';
 
 const Row = styled(Styled.Row)``;
